+++ parser.go
@@ -58,7 +58,7 @@ func parseFile(ctx context.Context, fd *descriptor.FileDescriptorProto) *FileDes
 	file := &FileDescriptor{
 		comments:            comments,
 		FileDescriptorProto: fd,
-		Comments:            comments[fmt.Sprintf("%d", packageCommentPath)],
+		Comments:            comments.Get(fmt.Sprintf("%d", packageCommentPath)),
 	}
 
 	fileCtx := ContextWithFileDescriptor(ctx, file)
@@ -87,7 +87,7 @@ func parseEnums(ctx context.Context, protos []*descriptor.EnumDescriptorProto) [
 		enums[i] = &EnumDescriptor{
 			common:              newCommon(file, commentPath, longName),
 			EnumDescriptorProto: ed,
-			Comments:            file.comments[commentPath],
+			Comments:            file.comments.Get(commentPath),
 			Parent:              parent,
 		}
 
@@ -110,7 +110,7 @@ func parseEnumValues(ctx context.Context, protos []*descriptor.EnumValueDescript
 			common: newCommon(file, "", longName),
 			EnumValueDescriptorProto: vd,
 			Enum:     enum,
-			Comments: file.comments[fmt.Sprintf("%s.%d.%d", enum.path, enumValueCommentPath, i)],
+			Comments: file.comments.Get(fmt.Sprintf("%s.%d.%d", enum.path, enumValueCommentPath, i)),
 		}
 	}
 
@@ -138,7 +138,7 @@ func parseExtensions(ctx context.Context, protos []*descriptor.FieldDescriptorPr
 		exts[i] = &ExtensionDescriptor{
 			common:               newCommon(file, commentPath, longName),
 			FieldDescriptorProto: ext,
-			Comments:             file.comments[commentPath],
+			Comments:             file.comments.Get(commentPath),
 			Parent:               parent,
 		}
 	}
@@ -186,7 +186,7 @@ func parseMessages(ctx context.Context, protos []*descriptor.DescriptorProto) []
 		msgs[i] = &Descriptor{
 			common:          newCommon(file, commentPath, longName),
 			DescriptorProto: md,
-			Comments:        file.comments[commentPath],
+			Comments:        file.comments.Get(commentPath),
 			Parent:          parent,
 		}
 
@@ -211,7 +211,7 @@ func parseMessageFields(ctx context.Context, protos []*descriptor.FieldDescripto
 		fields[i] = &FieldDescriptor{
 			common:               newCommon(file, "", longName),
 			FieldDescriptorProto: fd,
-			Comments:             file.comments[fmt.Sprintf("%s.%d.%d", message.path, messageFieldCommentPath, i)],
+			Comments:             file.comments.Get(fmt.Sprintf("%s.%d.%d", message.path, messageFieldCommentPath, i)),
 			Message:              message,
 		}
 	}
@@ -230,7 +230,7 @@ func parseServices(ctx context.Context, protos []*descriptor.ServiceDescriptorPr
 		svcs[i] = &ServiceDescriptor{
 			common:                 newCommon(file, commentPath, longName),
 			ServiceDescriptorProto: sd,
-			Comments:               file.comments[commentPath],
+			Comments:               file.comments.Get(commentPath),
 		}
 
 		svcCtx := ContextWithServiceDescriptor(ctx, svcs[i])
@@ -253,7 +253,7 @@ func parseServiceMethods(ctx context.Context, protos []*descriptor.MethodDescrip
 			common:                newCommon(file, "", longName),
 			MethodDescriptorProto: md,
 			Service:               svc,
-			Comments:              file.comments[fmt.Sprintf("%s.%d.%d", svc.path, serviceMethodCommentPath, i)],
+			Comments:              file.comments.Get(fmt.Sprintf("%s.%d.%d", svc.path, serviceMethodCommentPath, i)),
 		}
 	}